Splitting the codebase for easier testing
