+++ package.json
@@ -6,7 +6,8 @@
   "scripts": {
     "start": "node index.js",
     "start-dev": "NODE_ENV=dev node index.js",
-    "test": "node_modules/.bin/mocha -c --reporter mocha-junit-reporter --reporter-options mochaFile=./test/results/results.xml --recursive --exit test/"
+    "test": "node_modules/.bin/mocha --recursive --exit 'test/*.spec.js'",
+    "test-output": "node_modules/.bin/mocha -c --reporter mocha-junit-reporter --reporter-options mochaFile=./test/results/test-results.xml --recursive --exit 'test/*.spec.js'"
   },
   "repository": {
     "type": "git",
